requirements.txt
Django==2.2
django-crispy-forms==1.7.2
django-filter==2.1.0
djangorestframework